Set TEST_VERBOSE explicitly, as it controls the module's behavior
Having TEST_VERBOSE set means that the previous __WARN__ handler is
called when warnings occur in the child.  This means that t/destroy.t
failed when run with -v, as the initial inner warning was pushed onto
@warnings, in addition to the "Unhandled warning" warning.

Explicitly set $ENV{TEST_VERBOSE} in the test, so that the behavior of
the test file is consistent no matter how it is called.